requirements.txt
aiohttp==3.11.14
APScheduler==3.11.0
cleantext==1.1.4
dash==3.0.1
dash_bootstrap_components==2.0.2
dash_mantine_components==1.1.0
faiss_cpu==1.10.0
fastapi==0.115.12
langchain_together==0.3.0
numpy==2.2.5
plotly==6.0.1
python-dotenv==1.1.0
pytz==2025.2
Requests==2.32.3
sentence_transformers==4.0.1
SQLAlchemy==2.0.40
uvicorn==0.34.2
yfinance==0.2.55
